requirements.txt
pylint
streamlit
faiss-cpu
sentence-transformers
pypdf
ollama